+++ assets/sass/_core.scss
@@ -12,7 +12,7 @@ body {
   font-size: 12pt;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-color: $off-white;
+  background-color: $body-bg;
 }
 
 h1, h2, h3, h4, h5, h6 {
