+++ assets/sass/_footer.scss
@@ -5,6 +5,7 @@
 .footer {
   position: absolute;
   bottom: 0;
+  margin-top: 4rem;
   width: 100%;
   height: 60px;
 }
