+++ assets/sass/_header.scss
@@ -31,6 +31,11 @@
         color: $bold-yellow;
       }
     }
+    .brand-subtext {
+      display: block;
+      position: relative;
+      font-weight: 300;
+    }
     b {
       display: block;
       font-size: 50%;
