+++ assets/sass/_pages.scss
@@ -1,1 +1,2 @@
 @import "pages/home";
+@import "pages/repo_list";
