+++ assets/sass/components/_cards.scss
@@ -3,16 +3,12 @@
 // =============================================================================
 
 .card {
-  //box-shadow: 0 2px 2px 0 #E3E3E3;
-  //box-shadow: 0 0px 50px 0 rgba(34,43,55,.1);
-  box-shadow: 0 2px 20px 0 rgba(0,0,0,.1);
+  box-shadow: 0 0 0 1px rgba(61,119,180,.04), 0px 5px 22px 0 rgba(109,160,255,.16);
   position: relative;
   background-color: $white;
-  border-radius: 1rem !important;
-  border: 0 !important;
+  border-radius: .8rem;
   break-inside: avoid;
-  padding: 10px 10px 35px;
-  border-bottom: none;
+  padding: 1rem 1.5rem;
   display: block;
   width: 100%;
   margin: 1rem auto;
@@ -23,29 +19,67 @@
   &-body {
     padding: 2.25rem !important;
   }
-  &-image {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-    img {
-      max-width: unset;
+
+  &-title {
+    &:hover {
+      text-decoration: none;
+    }
+
+    h1 {
+      margin-bottom: 0 !important;
+      color: $bold-blue;
+      text-transform: uppercase;
+      font-weight: 900;
+    }
+    &:hover {
+      color: $bold-blue;
+      &::after {
+        transform: scaleX(1);
+        border-color: $bold-blue;
+      }
+    }
+    &::after {
+      content: '';
+      display: block;
       position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%,-50%);
-      width: auto;
-      height: 100%;
+      width: 40%;
+      border-bottom: solid 2px $white;
+      -webkit-transform: scaleX(0);
+      -moz-transform: scaleX(0);
+      -ms-transform: scaleX(0);
+      -o-transform: scaleX(0);
+      transform: scaleX(0);
+      transition: all $ease-slow;
     }
   }
 
-  &-title {
-    margin-bottom: 0 !important;
+  &-author,
+  &-stars,
+  &-description {
+    font-weight: 400;
+    color: $grey;
   }
 
-  .list-group-item {
-    p {
-      margin-bottom: 0;
+  &-author {
+
+    &.name {
+      font-weight: 900;
+      color: $black;
+      text-transform: capitalize;
+    }
+  }
+
+  &-stars {
+    &.starsVal {
+      font-weight: bold;
+      color: darken($bold-yellow, 15);
+    }
+  }
+
+  &-description {
+    &.descriptionVal {
+      font-weight: bold;
+      color: $black;
     }
   }
