+++ assets/sass/pages/_repo_list.scss
@@ -1,0 +1,15 @@
+.repo_list {
+  animation: repoIn 1s forwards;
+  transition: all $ease;
+}
+
+@keyframes repoIn {
+  0% {
+    opacity: 0;
+    transform: translateY(5rem);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0rem);
+  }
+}
