+++ assets/sass/utilities/_transitions.scss
@@ -14,7 +14,7 @@
 @keyframes fadeUp {
   0% {
     opacity: 0;
-    transform: translate3d(0, -3.5%, 0);
+    transform: translate3d(0, 3.5%, 0);
   }
   100% {
     opacity: 1;
